mobile-200-handleImageUpload-fix: refactored image upload and file rename pieces into imageUpload helper function
